Fixed compilation errors, project setup
