+++ .gitignore
@@ -41,3 +41,5 @@ $RECYCLE.BIN/
 Network Trash Folder
 Temporary Items
 .apdisk
+/build/
+/dist/
