+++ nbproject/private/private.properties
@@ -1,2 +1,6 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\Joe's Minecraft\\AppData\\Roaming\\NetBeans\\8.0.2\\build.properties
+do.depend=false
+do.jar=true
+javac.debug=true
+javadoc.preview=true
+user.properties.file=C:\\Users\\Jerom\\AppData\\Roaming\\NetBeans\\8.0.2\\build.properties
