+++ nbproject/project.properties
@@ -1,9 +1,11 @@
 annotation.processing.enabled=true
 annotation.processing.enabled.in.editor=false
-annotation.processing.processor.options=
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
+application.title=ScuphsUtilities
+application.vendor=Scuph
+auxiliary.org-netbeans-spi-editor-hints-projects.perProjectHintSettingsFile=nbproject/cfg_hints.xml
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
@@ -26,27 +28,22 @@ dist.archive.excludes=
 dist.dir=dist
 dist.jar=${dist.dir}/ScuphsUtilities.jar
 dist.javadoc.dir=${dist.dir}/javadoc
+endorsed.classpath=
 excludes=
-file.reference.BukkitLib-1.3-javadoc.jar=C:\\Users\\Joe's Minecraft\\Desktop\\BukkitLib-1.3-javadoc.jar
-file.reference.BukkitLib-1.3.jar=C:\\Users\\Joe's Minecraft\\Desktop\\Everything\\BukkitLib-1.3.jar
-file.reference.BukkitLib-1.3__1_.jar=C:\\Users\\Joe's Minecraft\\Desktop\\BukkitLib-1.3 (1).jar
-file.reference.spigot.jar=C:\\Users\\Joe's Minecraft\\Desktop\\Minecraft 1.7.10 Server (64Bits)\\spigot.jar
 includes=**
+jar.archive.disabled=${jnlp.enabled}
 jar.compress=false
+jar.index=${jnlp.enabled}
 javac.classpath=\
-    ${libs.BukktLib.classpath}:\
-    ${libs.Spigot.classpath}:\
-    ${file.reference.BukkitLib-1.3__1_.jar}:\
-    ${file.reference.BukkitLib-1.3-javadoc.jar}:\
-    ${file.reference.spigot.jar}:\
-    ${file.reference.BukkitLib-1.3.jar}
+    ${libs.SpigotAPI.classpath}:\
+    ${libs.BukkitLib.classpath}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
 javac.processorpath=\
     ${javac.classpath}
-javac.source=1.8
-javac.target=1.8
+javac.source=1.7
+javac.target=1.7
 javac.test.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
@@ -63,7 +60,20 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
+jnlp.codebase.type=no.codebase
+jnlp.descriptor=application
+jnlp.enabled=false
+jnlp.mixed.code=default
+jnlp.offline-allowed=false
+jnlp.signed=false
+jnlp.signing=
+jnlp.signing.alias=
+jnlp.signing.keystore=
 main.class=
+# Optional override of default Codebase manifest attribute, use to prevent RIAs from being repurposed
+manifest.custom.codebase=
+# Optional override of default Permissions manifest attribute (supported values: sandbox, all-permissions)
+manifest.custom.permissions=
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
