+++ src/com/scuph/scuphsutilities/commands/Command_admins.java
@@ -1,6 +1,6 @@
 package com.scuph.scuphsutilities.commands;
 
-import com.scuph.scuphsutilities.ScuphsUtilities;
+import com.scuph.scuphsutilities.ScuphUtils;
 import net.pravian.bukkitlib.command.BukkitCommand;
 import net.pravian.bukkitlib.command.CommandPermissions;
 import net.pravian.bukkitlib.command.SourceType;
@@ -9,12 +9,12 @@ import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 
 @CommandPermissions(source = SourceType.ANY, permission = "utils.admins")
-public class Command_admins extends BukkitCommand<ScuphsUtilities> {   
+public class Command_admins extends BukkitCommand<ScuphUtils> {
 
     @Override
     public boolean run(CommandSender sender, Command cmd, String commandLabel, String[] args) {
         sender.sendMessage(ChatColor.RED + plugin.getConfig().getString("admins"));
         return true;
     }
-    
+
 }
