+++ src/com/scuph/scuphsutilities/commands/Command_information.java
@@ -1,17 +1,16 @@
 package com.scuph.scuphsutilities.commands;
 
-import com.scuph.scuphsutilities.ScuphsUtilities;
+import com.scuph.scuphsutilities.ScuphUtils;
 import net.pravian.bukkitlib.command.BukkitCommand;
 import net.pravian.bukkitlib.command.CommandPermissions;
 import net.pravian.bukkitlib.command.SourceType;
-import net.pravian.bukkitlib.config.YamlConfig;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 
 @CommandPermissions(source = SourceType.ANY, permission = "utils.info")
-public class Command_information extends BukkitCommand<ScuphsUtilities> {
-    
+public class Command_information extends BukkitCommand<ScuphUtils> {
+
     @Override
     public boolean run(CommandSender sender, Command cmd, String commandLabel, String[] args) {
         sender.sendMessage(ChatColor.GREEN + "This plugin was made by ScuphGamingUK, made to set the core of a server! But keep in mind: he is not part of " + plugin.getConfig().getString("server-name") + "'s staff!");
@@ -22,5 +21,5 @@ public class Command_information extends BukkitCommand<ScuphsUtilities> {
         sender.sendMessage(ChatColor.DARK_RED + "The server's rules can be found by doing " + plugin.getConfig().getString("rule-command") + "!");
         return true;
     }
-    
+
 }
