+++ src/com/scuph/scuphsutilities/commands/Command_kick.java
@@ -1,11 +1,11 @@
 package com.scuph.scuphsutilities.commands;
 
-import com.scuph.scuphsutilities.ScuphsUtilities;
+import com.scuph.scuphsutilities.ScuphUtils;
 import net.pravian.bukkitlib.command.BukkitCommand;
 import net.pravian.bukkitlib.command.CommandPermissions;
 import net.pravian.bukkitlib.command.SourceType;
-import org.apache.commons.lang3.ArrayUtils;
-import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang.ArrayUtils;
+import org.apache.commons.lang.StringUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
@@ -13,29 +13,29 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
 @CommandPermissions(source = SourceType.PLAYER, permission = "utils.kick")
-public class Command_kick extends BukkitCommand<ScuphsUtilities> {   
+public class Command_kick extends BukkitCommand<ScuphUtils> {
 
     @Override
     public boolean run(CommandSender sender, Command cmd, String commandLabel, String[] args) {
         if (args.length < 3) {
             return showUsage();
         }
-        
+
         final Player player = getPlayer(args[0]);
-        
+
         if (player == null) {
             sender.sendMessage(ChatColor.RED + "Player not found.\nPlease review arguments.");
         }
-        
+
         String kickReason = StringUtils.join(ArrayUtils.subarray(args, 1, args.length), " ");
-        
+
         Bukkit.broadcastMessage(ChatColor.RED + sender.getName() + " - Kicking: " + player.getName() + " Reason: " + kickReason);
-        
+
         player.kickPlayer(ChatColor.RED + "You have been kicked\nReason: " + ChatColor.YELLOW + kickReason + ChatColor.RED + "\nKicked by ~ " + ChatColor.YELLOW + sender.getName());
-        
+
         sender.sendMessage(ChatColor.GRAY + "Player successfully kicked");
-        
+
         return true;
     }
-    
+
 }
